Switch to PEP-517 compliant build system.
This patch allows building a standard Python wheel out of pyrmexplorer.
It is currently using flit as a build backend, a minimal build system
often used in Python projects but this is only a proposal.

Any PEP 517 and PEP 621 compliant tool can parse the provided pyproject.toml,
provided the build-backed field is adjusted.